Convert geographic ('sf') representation of OD data to 'JSON list' structure

This function takes outputs from ab_scenario() and returns a list that can be saved as a JSON file for import into A/B Street.

ab_json(
  desire_lines_out,
  mode_column = NULL,
  time_fun = ab_time_normal,
  scenario_name,
  ...
)

Arguments

desire_lines_out
OD data represented as geographic lines created by ab_scenario().

mode_column
The column name in the desire lines data that contains the mode of transport. "mode_baseline" by default.

time_fun
The function used to calculate departure times. ab_time_normal() by default.

scenario_name
The name of the scenario to appear in A/B Street

...
Additional arguments to pass to ab_json()

Value

A list that can be saved as a JSON file with ab_save()

Details

Note: the departure time in seconds is multiplied by 10000 on conversion to a .json list object for compatibility with the A/B Street schema.

Examples

od = leeds_od
od[[1]] = c("E02006876")
zones = leeds_zones
ablines = ab_scenario(od, zones = zones)

#> 0 origins with no match in zone ids

#> 0 destinations with no match in zone ids

#>  points not in od data removed.

#> Linking to GEOS 3.9.1, GDAL 3.3.0, PROJ 8.0.1

ab_list = ab_json(ablines, mode_column = "mode", scenario_name = "test")
ab_list$scenario

#> [1] "test"

f = tempfile(fileext = ".json")
ab_save(ab_list, f)
readLines(f)[1:30]

#>  [1] "{"                                    
#>  [2] "  \"scenario_name\": \"test\","       
#>  [3] "  \"people\": ["                      
#>  [4] "    {"                                
#>  [5] "      \"trips\": ["                   
#>  [6] "        {"                            
#>  [7] "          \"departure\": 315260000,"  
#>  [8] "          \"origin\": {"              
#>  [9] "            \"Position\": {"          
#> [10] "              \"longitude\": -1.5225,"
#> [11] "              \"latitude\": 53.7838"  
#> [12] "            }"                        
#> [13] "          },"                         
#> [14] "          \"destination\": {"         
#> [15] "            \"Position\": {"          
#> [16] "              \"longitude\": -1.5654,"
#> [17] "              \"latitude\": 53.8061"  
#> [18] "            }"                        
#> [19] "          },"                         
#> [20] "          \"mode\": \"Walk\","        
#> [21] "          \"purpose\": \"Shopping\""  
#> [22] "        }"                            
#> [23] "      ]"                              
#> [24] "    },"                               
#> [25] "    {"                                
#> [26] "      \"trips\": ["                   
#> [27] "        {"                            
#> [28] "          \"departure\": 278840000,"  
#> [29] "          \"origin\": {"              
#> [30] "            \"Position\": {"          

# Legacy code from ActDev project commented out
# ab_list$people$trips[[1]]
# dutch = ab_scenario(
#   leeds_houses,
#   leeds_buildings,
#   leeds_desire_lines,
#   leeds_zones,
#   scenario = "godutch",
#   output = "sf"
# )
# ab_list = ab_json(dutch, mode_column = "mode_godutch")
# ab_list$scenario
# str(ab_list$people$trips[[9]])
# # add times
# dutch$departure = ab_time_normal(hr = 1, sd = 0, n = nrow(dutch))
# ab_list_times = ab_json(dutch)
# str(ab_list_times$people$trips[[9]])
# ab_save(ab_list_times, f)
# readLines(f)[1:30]
# 60^2