Example Activity data for São Paulo (20 agents)

Each row of this table contains a single trip of 20 different agents in the São Paulo city.

sao_paulo_activity_df_20

Format

A data frame with columns:

ZONA

Household zone.

MUNI_DOM

Household municipality.

CO_DOM_X

Household coordinate X.

CO_DOM_Y

Household coordinate Y.

ID_DOM

Household identifier.

F_DOM

Identify first household record.

FE_DOM

Household expansion factor.

DOM

Household number.

CD_ENTRE

Interview code.

DATA

Date of interview.

TIPO_DOM

Type of household.

AGUA

Do you have piped water?.

RUA_PAVI

Is the street paved?.

NO_MORAD

Household size.

TOT_FAM

Total families in the household.

ID_FAM

Family identifier.

F_FAM

Identify first family record.

FE_FAM

Famiy expansion factor.

FAMILIA

Family number

NO_MORAF

Family size.

CONDMORA

Housing condition.

QT_BANHO

Bathrooms.

QT_EMPRE

Domestic workers.

QT_AUTO

Automobiles.

QT_MICRO

Microcomputers.

QT_LAVALOU

Dishwashers.

QT_GEL1

One door refrigerator.

QT_GEL2

Two door refrigerator.

QT_FREEZ

Frezers.

QT_MLAVA

Washing machines.

QT_DVD

DVDs.

QT_MICROON

Microwave oven.

QT_MOTO

Motorcycles.

QT_SECAROU

Clothes dryer.

QT_BICICLE

Bikes.

NAO_DCL_IT

Comfort item declaration code.

CRITERIOBR

Critério de Classificação Econômica Brasil.

PONTO_BR

Critério Brasil points.

ANO_AUTO1

Year of manufacture of Auto 1.

ANO_AUTO2

Year of manufacture of Auto 2.

ANO_AUTO3

Year of manufacture of Auto 3.

RENDA_FA

Monthly family income.

CD_RENFA

Monthly family income code.

ID_PESS

Person identifier.

F_PESS

Identifies first person record.

FE_PESS

Person expansion factor.

PESSOA

Person number.

SIT_FAM

Family situation.

IDADE

Age.

SEXO

Sex.

ESTUDA

Are you currently studying?

GRAU_INS

Education.

CD_ATIVI

Labor market status.

CO_REN_I

Individual income condition.

VL_REN_I

Individual income.

ZONA_ESC

School zone.

MUNIESC

School municipality.

CO_ESC_X

School coordinate X.

CO_ESC_Y

School coordinate Y.

TIPO_ESC

Type of school.

ZONATRA1

Zone of 1st job.

MUNITRA1

Municipality of 1st job.

CO_TR1_X

1st job coordinate X.

CO_TR1_Y

1st job coordinate Y.

TRAB1_RE

Does 1st job equal residency?

TRABEXT1

Performs outside work - 1st job.

OCUP1

1st job occupation.

SETOR1

1st job activity sector.

VINC1

1st job employment relationship

ZONATRA2

Zone of the 2nd job.

MUNITRA2

Municipality of the 2nd job.

CO_TR2_X

2nd job coordinate X.

CO_TR2_Y

2nd job coordinate Y.

TRAB2_RE

Does 2nd job equal residency?

TRABEXT2

Performs outside work - 2nd job.

OCUP2

2nd job occupation.

SETOR2

2nd job activity sector.

VINC2

2nd job employment relationship.

N_VIAG

Trip number.

FE_VIA

Trip expansion factor.

DIA_SEM

Day of the week.

TOT_VIAG

Total person's trip.

ZONA_O

Zone of origin.

MUNI_O

Municipality of origin.

CO_O_X

Origin coordinate X.

CO_O_Y

Origin coordinate Y.

ZONA_D

Zone of destination.

MUNI_D

Municipality of destination.

CO_D_X

Destination coordinate X.

CO_D_Y

Destination coordinate Y.

ZONA_T1

Zone of the 1st transfer.

MUNI_T1

Municipality of the 1st transfer.

CO_T1_X

1st transfer coordinate X.

CO_T1_Y

1st transfer coordinate Y.

ZONA_T2

Zone of the 2nd transfer.

MUNI_T2

Municipality of the 2nd transfer.

CO_T2_X

2nd transfer coordinate X.

CO_T2_Y

2nd transfer coordinate Y.

ZONA_T3

Zone of the third transfer.

MUNI_T3

Municipality of the 3rd transfer.

CO_T3_X

3rd transfer coordinate X.

CO_T3_Y

3rd transfer coordinate Y.

MOTIVO_O

Reason at origin.

MOTIVO_D

Reason at destination.

MOT_SRES

Reason at destination without residence.

SERVIR_O

Serving passenger at origin.

SERVIR_D

Serving passenger at destination.

MODO1

Mode 1.

MODO2

Mode 2.

MODO3

Mode 3.

MODO4

Mode 4.

H_SAIDA

Departure hour.

MIN_SAIDA

Departure minute.

ANDA_O

Time walking in origin.

H_CHEG

Arrival hour.

MIN_CHEG

Arrival minute.

ANDA_D

Time walking in destination.

DURACAO

Duration.

MODOPRIN

Main mode.

TIPVG

Type of trip.

PAG_VIAG

Who paid for the trip.

TP_ESAUTO

Type of car or motorcycle parking.

VL_EST

Car or motorcycle parking value.

PE_BICI

Why did you travel by foot or bike?

VIA_BICI

If you traveled by bicycle, did you use the segregated route?

TP_ESBICI

Bike parking.

DISTANCIA

Distance (meters).

ID_ORDEM

Registration order number.

Details

See the code used to create this data in "data-raw/sao-paulo-activity-data.R"