Example Activity data for São Paulo

Each row of this table contains a single trip of people in the São Paulo city. sao_paulo_activity_df_2 represents the movement of 2 people, sao_paulo_activity_df_20 represents the movement of 20 people.

sao_paulo_activity_df_20

Format

A data frame with columns:

ID_PESS

Person identifier.

CO_O_X

Origin coordinate X.

CO_O_Y

Origin coordinate Y.

CO_D_X

Destination coordinate X.

CO_D_Y

Destination coordinate Y.

MODOPRIN

Main mode.

H_SAIDA

Departure hour.

MIN_SAIDA

Departure minute.

Details

See the code used to create this data, and the full open dataset with 128 variables, the file data-raw/sao-paulo-activity-data.R in the package's GitHub repo.

Examples

dim(sao_paulo_activity_df_20)
#> [1] 88  9
names(sao_paulo_activity_df_20)
#> [1] "ID_PESS"   "CO_O_X"    "CO_O_Y"    "CO_D_X"    "CO_D_Y"    "MODOPRIN" 
#> [7] "MOTIVO_O"  "H_SAIDA"   "MIN_SAIDA"
head(sao_paulo_activity_df_20)
#> # A tibble: 6 × 9
#>   ID_PESS     CO_O_X  CO_O_Y CO_D_X  CO_D_Y MODOPRIN MOTIVO_O H_SAIDA MIN_SAIDA
#>   <chr>        <int>   <int>  <int>   <int>    <int>    <int>   <int>     <int>
#> 1 00030710102 333614 7394212 333611 7394202       16        8       8         0
#> 2 00030710102 333611 7394202 333614 7394212       16        1      13         0
#> 3 00030710102 333614 7394212 333611 7394202       16        8      13        40
#> 4 00030710102 333611 7394202 333614 7394212       16        1      17         0
#> 5 00030743103 333754 7394044 333759 7394160       16        8       6        55
#> 6 00030743103 333759 7394160 333754 7394044       16        4       7         5
dim(sao_paulo_activity_df_2)
#> [1] 6 9
names(sao_paulo_activity_df_2)
#> [1] "ID_PESS"   "CO_O_X"    "CO_O_Y"    "CO_D_X"    "CO_D_Y"    "MODOPRIN" 
#> [7] "MOTIVO_O"  "H_SAIDA"   "MIN_SAIDA"
head(sao_paulo_activity_df_2)
#> # A tibble: 6 × 9
#>   ID_PESS     CO_O_X  CO_O_Y CO_D_X  CO_D_Y MODOPRIN MOTIVO_O H_SAIDA MIN_SAIDA
#>   <chr>        <int>   <int>  <int>   <int>    <int>    <int>   <int>     <int>
#> 1 00240507101 333454 7393635 333221 7394620        4        8       8        30
#> 2 00240507101 333221 7394620 332276 7395195       16        5       9        20
#> 3 00240507101 332276 7395195 333454 7393635        4        3      17        30
#> 4 00241455101 333330 7393338 333374 7395289        1        8      11         0
#> 5 00241455101 333374 7395289 333143 7393727        1        5      12        40
#> 6 00241455101 333143 7393727 333330 7393338       16        7      14        30