Example Activity data for São Paulo

Each row of this table contains a single trip of people in the São Paulo city. sao_paulo_activity_sf_2 represents the movement of 2 people, sao_paulo_activity_sf_20 represents the movement of 20 people.

sao_paulo_activity_sf_20

Format

A data frame with columns:

person

Person identifier.

departure

Departure time in seconds past midnight

mode

Mode of travel in A/B Street terms

purpose

Purpose of travel in A/B Street terms

geometry

Geometry of the linestring representing the OD pair

Details

See the code used to create this data, and the full open dataset with 128 variables, the file data-raw/sao-paulo-activity-data.R in the package's GitHub repo.

Examples

dim(sao_paulo_activity_sf_20)

#> [1] 88  5

names(sao_paulo_activity_sf_20)

#> [1] "person"    "departure" "mode"      "purpose"   "geometry" 

head(sao_paulo_activity_sf_20)

#> Simple feature collection with 6 features and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -46.63044 ymin: -23.55553 xmax: -46.629 ymax: -23.554
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 5
#>   person      departure mode  purpose                                 geometry
#>   <chr>           <dbl> <chr> <chr>                           <LINESTRING [°]>
#> 1 00030710102     28800 Walk  Home    (-46.63041 -23.554, -46.63044 -23.55409)
#> 2 00030710102     46800 Walk  Work    (-46.63044 -23.55409, -46.63041 -23.554)
#> 3 00030710102     49200 Walk  Home    (-46.63041 -23.554, -46.63044 -23.55409)
#> 4 00030710102     61200 Walk  Work    (-46.63044 -23.55409, -46.63041 -23.554)
#> 5 00030743103     24900 Walk  Home    (-46.62906 -23.55553, -46.629 -23.55449)
#> 6 00030743103     25500 Walk  School  (-46.629 -23.55449, -46.62906 -23.55553)

table(sao_paulo_activity_sf_20$mode)

#> 
#>   Drive Transit    Walk 
#>      12       4      72 

table(sao_paulo_activity_sf_20$purpose)

#> 
#>       Home       Meal Recreation     School   Shopping       Work 
#>         39          2          3         14          2         28 

dim(sao_paulo_activity_sf_2)

#> [1] 6 5

names(sao_paulo_activity_sf_2)

#> [1] "person"    "departure" "mode"      "purpose"   "geometry" 

head(sao_paulo_activity_sf_2)

#> Simple feature collection with 6 features and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -46.64341 ymin: -23.56186 xmax: -46.63204 ymax: -23.54425
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 5
#>   person      departure mode    purpose                                 geometry
#>   <chr>           <dbl> <chr>   <chr>                           <LINESTRING [°]>
#> 1 00240507101     30600 Transit Home       (-46.63204 -23.5592, -46.63422 -23.5…
#> 2 00240507101     33600 Walk    Shopping   (-46.63422 -23.55028, -46.64341 -23.…
#> 3 00240507101     63000 Transit Work       (-46.64341 -23.54499, -46.63204 -23.…
#> 4 00241455101     39600 Transit Home       (-46.63329 -23.56186, -46.63264 -23.…
#> 5 00241455101     45600 Transit Shopping   (-46.63264 -23.54425, -46.63508 -23.…
#> 6 00241455101     52200 Walk    Recreation (-46.63508 -23.55833, -46.63329 -23.…